Issue #723 Getting rid of multiple CacheManager impl
