+++ clustered/client/src/main/java/org/ehcache/clustered/service/DefaultClusteringService.java
@@ -17,7 +17,15 @@
 package org.ehcache.clustered.service;
 
 import org.ehcache.clustered.config.ClusteringServiceConfiguration;
+import org.ehcache.config.ResourcePool;
+import org.ehcache.config.ResourceType;
+import org.ehcache.exceptions.CachePersistenceException;
 import org.ehcache.spi.ServiceProvider;
+import org.ehcache.spi.service.ServiceConfiguration;
+
+import java.util.Collection;
+
+import static java.util.Collections.emptyList;
 
 /**
  * @author Clifford W. Johnson
@@ -38,5 +46,27 @@ public class DefaultClusteringService implements ClusteringService {
   @Override
   public void stop() {
     // TODO: Implement stop
+  }
+
+  @Override
+  public boolean handlesResourceType(ResourceType resourceType) {
+    // TODO: implement me
+    return false;
+  }
+
+  @Override
+  public Collection<ServiceConfiguration<?>> additionalConfigurationsForPool(String alias, ResourcePool pool) throws CachePersistenceException {
+    // TODO: implement me
+    return emptyList();
+  }
+
+  @Override
+  public void destroyPersistenceSpace(String name) throws CachePersistenceException {
+    // TODO: implement me
+  }
+
+  @Override
+  public void destroyAllPersistenceSpaces() {
+    // TODO: implement me
   }
 }
