+++ impl/src/test/java/org/ehcache/impl/internal/persistence/TestLocalPersistenceService.java
@@ -17,12 +17,17 @@
 package org.ehcache.impl.internal.persistence;
 
 import java.io.File;
+import java.util.Collection;
+
+import org.ehcache.config.ResourcePool;
+import org.ehcache.config.ResourceType;
 import org.ehcache.impl.config.persistence.CacheManagerPersistenceConfiguration;
 import org.ehcache.exceptions.CachePersistenceException;
 import org.ehcache.impl.persistence.DefaultLocalPersistenceService;
 import org.ehcache.spi.ServiceProvider;
 import org.ehcache.core.spi.service.FileBasedPersistenceContext;
 import org.ehcache.core.spi.service.LocalPersistenceService;
+import org.ehcache.spi.service.ServiceConfiguration;
 import org.junit.rules.ExternalResource;
 import org.junit.rules.TemporaryFolder;
 
@@ -61,6 +66,16 @@ public class TestLocalPersistenceService extends ExternalResource implements Loc
     } finally {
       folder.delete();
     }
+  }
+
+  @Override
+  public boolean handlesResourceType(ResourceType resourceType) {
+    return persistenceService.handlesResourceType(resourceType);
+  }
+
+  @Override
+  public Collection<ServiceConfiguration<?>> additionalConfigurationsForPool(String alias, ResourcePool pool) throws CachePersistenceException {
+    return persistenceService.additionalConfigurationsForPool(alias, pool);
   }
 
   @Override
